refactor: medium-priority structure for #1531 (parts, barrels, map dispatch)
- Split order_suggestion and game_setup into library parts; extract tile map
  generator types; region map shared part for constants/helpers/enums.
- Add turn/phases.dart and orders/order_validators.dart barrels; wire
  turn_phase_runner and order_engine.
- Replace work-order visibility and work-completion switches with map dispatch;
  extract explorer work suggestions into helpers; use kWorkTarget* for explore/
  prospect suggestions.
- Document map-dispatch structure in SPEC/program/orders.md.
- Tests: kWorkTargetExplore visibility + suggestWorkOrders; existing suites pass.

Refs #1531

Made-with: Cursor
