+++ packages/colonizethis_logic/lib/src/orders/order_validators.dart
@@ -1,0 +1,9 @@
+/// Barrel: order validation modules (import reduction). Refs #1531.
+library;
+
+export 'validators/army_move_validator.dart';
+export 'validators/build_order_validator.dart';
+export 'validators/diplomatic_order_validator.dart';
+export 'validators/move_validator.dart';
+export 'validators/naval_order_validator.dart';
+export 'validators/work_order_validator.dart';
