+++ packages/colonizethis_logic/lib/src/turn/phases.dart
@@ -1,0 +1,12 @@
+/// Barrel: turn phase step implementations (import reduction). Refs #1531.
+library;
+
+export 'phases/build_work_phase.dart';
+export 'phases/combat_phase.dart';
+export 'phases/consumption_phase.dart';
+export 'phases/extraction_phase.dart';
+export 'phases/minor_regiment_upgrade_phase.dart';
+export 'phases/movement_phase.dart';
+export 'phases/naval_interception_turn_phase.dart';
+export 'phases/production_phase.dart';
+export 'phases/riches_to_treasury_phase.dart';
